ai_service/requirements.txt
fastapi>=0.116,<1.0
Pillow>=11.0,<12.0
python-multipart>=0.0.9,<1.0
PyYAML>=6.0,<7.0
torch>=2.6,<3.0
torchvision>=0.21,<1.0
uvicorn>=0.35,<1.0
